passing the user id on the frontend

+++ backend/routes/posts.js
@@ -71,9 +71,12 @@ router.put(
       content: req.body.content,
       imagePath: imagePath
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if(result.nModified > 0) {
+        res.status(200).json({ message: "update successful" });
+      } else {
+        res.status(401).json({ message: "not autohrized"});
+      }
     });
   }
 );
@@ -111,9 +114,13 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+    if(result.n > 0) {
+      res.status(200).json({ message: "Deletion successful" });
+    } else {
+      res.status(401).json({ message: "not autohrized"});
+    }
   });
 });
